+++ src/mantle/support/attributes/class-filter.php
@@ -14,7 +14,7 @@ use Attribute;
  *
  * Used to hook a method to an WordPress hook at a specific priority.
  */
-#[Attribute]
+#[Attribute( Attribute::IS_REPEATABLE | Attribute::TARGET_METHOD | Attribute::TARGET_FUNCTION )]
 class Filter {
 	/**
 	 * Constructor.